Modify test file structure
